[FE] MOVE: LoadingSpinner 위치 이동 #127
