+++ src/frontend/src/components/common/LoadingSpinner/index.css.tsx
